backend/requirements.txt
# FastAPI Backend Requirements

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
python-multipart>=0.0.6
email-validator>=2.0.0

# CORS
fastapi-cors>=0.0.6
